Crate async_fence

async_fence provies no_std async fences and OnceLock.

Feature Flags

alloc: Enables dynamic queue types via alloc.

Licensing and Contributing

All code is licensed under MPL 2.0. See the FAQ for license questions. The license is non-viral copyleft and does not block this library from being used in closed-source codebases. If you are using this library for a commercial purpose, consider reaching out to dansecob.dev@gmail.com to make a financial contribution.

Contributions are welcome at https://github.com/Bennett-Petzold/async_fence.